AbstractLazyObject3D
