+++ projects/atft/src/lib/objects/loaders/abstract-model-loader.ts
@@ -1,42 +1,14 @@
-import {Input, OnDestroy} from '@angular/core';
-
-import {AbstractObject3D} from '../abstract-object-3d';
-
-import * as THREE from 'three';
+import {Input} from '@angular/core';
+import {AbstractLazyObject3D} from '../abstract-lazy-object-3d';
 
 /**
  * Helper parent class for model loader.
  *
  * @see ObjLoaderComponent
  */
-export abstract class AbstractModelLoader extends AbstractObject3D<THREE.Object3D> implements OnDestroy {
+export abstract class AbstractModelLoader extends AbstractLazyObject3D {
 
   private _model: string;
-
-  /**
-   * Flag to signal whether the parent class instance AbstractObject3D called the
-   * overwritten method {@link AbstractModelLoader#afterInit} yet.
-   *
-   * Unless that method was called, no methods and properties of {@link AbstractObject3D}
-   * may be safely accessed, especially {@link AbstractObject3D#addChild} and
-   * {@link AbstractObject3D.renderer}.
-   */
-  private parentInitialized = false;
-
-  protected currentLoadedModelObject: THREE.Object3D | undefined;
-
-  /**
-   * Load the model object.
-   *
-   * Some loaders (e.g. ColladaLoader) also provide other model information
-   * upon loading besides the "raw" model object/scene. In these cases
-   * implementing child classes are indeed supposed to return the "raw" model
-   * object.
-   * The data source (usually a URI, although child classes are free to implement
-   * other means as well) from which the model shall be loaded can be obtained by
-   * {@link AbstractModelLoader.model}.
-   */
-  protected abstract async loadModelObject(): Promise<THREE.Object3D>;
 
   /**
    * The model data source (usually a URI).
@@ -46,26 +18,10 @@ export abstract class AbstractModelLoader extends AbstractObject3D<THREE.Object3
    */
   @Input()
   public set model(newModelUrl: string) {
-    this._model = newModelUrl;
-
-    // Delay model loading until the parent has been initialized,
-    // so that we can call addChild().
-    if (!this.parentInitialized) {
-      return;
+    if (this._model !== newModelUrl) {
+      this._model = newModelUrl;
+      super.startLoading();
     }
-
-    this.loadModelObject().then(newModel => {
-      if (this.currentLoadedModelObject) {
-        this.removeChild(this.currentLoadedModelObject);
-      }
-
-      this.currentLoadedModelObject = newModel;
-      this.addChild(newModel);
-
-      this.render.emit();
-    }).catch(err => {
-      console.error(err);
-    });
   }
 
   /**
@@ -75,21 +31,4 @@ export abstract class AbstractModelLoader extends AbstractObject3D<THREE.Object3
     return this._model;
   }
 
-  protected afterInit() {
-    super.afterInit();
-    this.parentInitialized = true;
-
-    // Trigger model acquisition now that the parent has been initialized.
-    this.model = this.model;
-  }
-
-  ngOnDestroy(): void {
-    if (this.currentLoadedModelObject) {
-      this.removeChild(this.currentLoadedModelObject);
-    }
-  }
-
-  protected newObject3DInstance(): THREE.Object3D {
-    return new THREE.Object3D();
-  }
 }
